Split source DB file queries
