+++ crates/wavecrate-library/src/sample_sources/db/read/file_queries/sql.rs
@@ -1,0 +1,3 @@
+pub(super) fn supported_audio_filter() -> String {
+    crate::sample_sources::supported_audio_where_clause()
+}
